Split import lowering planner
